requirements.txt
flask==2.1.0
python-dotenv==0.21.0
PyMySQL==1.0.2
gunicorn==20.1.0
wo=0.1.5